python-app/pyproject.toml
[project]
name = "sensus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "chromadb>=0.6.3",
    "exa-py>=1.10.0",
    "faker>=37.1.0",
    "fastapi>=0.115.12",
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "google-genai>=1.8.0",
    "langchain>=0.3.21",
    "langchain-community>=0.3.20",
    "langchain-core>=0.3.49",
    "langchain-exa>=0.2.1",
    "langchain-google-genai>=2.1.2",
    "langchain-nvidia-ai-endpoints>=0.3.9",
    "langgraph>=0.3.22",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "python-multipart>=0.0.20",
    "sec-edgar-downloader>=5.0.3",
    "slowapi>=0.1.9",
    "tavily-python>=0.5.3",
    "uvicorn>=0.34.0",
]

[tool.uv.workspace]
members = ["ticket-system"]
